Skip tests that will fail in January
Some tests were failing in January. We have a variable called
`previousMonth` in a few spots, which is set to `currentMonth - 1`
unless it's January, when it's set to `1`. We have some tests that
basically test `${previousMonth} / ${currentYear}`, which will fail when
it's January.

This commit fixes that by skipping those tests in January. (It also
removes a duplicate test.)

I think the right answer would be to stub some `Date` methods so it
didn't matter when these tests were ran, but this fixes things for now.